requirements.txt
finnhub-python==1.0.3
pandas==1.1.5
seaborn==0.11.1